+++ packages/color-modes/test/index.tsx
@@ -66,7 +66,9 @@ test('renders with initial color mode name', () => {
     renderer.create(
       <ThemeProvider
         theme={{
-          initialColorModeName: 'light',
+          config: {
+            initialColorModeName: 'light',
+          },
           colors: {
             modes: {
               dark: {},
@@ -128,7 +130,9 @@ test('color mode is passed through theme context', () => {
   const tree = render(
     <ThemeProvider
       theme={{
-        useCustomProperties: false,
+        config: {
+          useCustomProperties: false,
+        },
         colors: {
           text: '#000',
           modes: {
@@ -230,7 +234,9 @@ test('does not initialize mode based on localStorage if useLocalStorage is set t
   const tree = render(
     <ThemeProvider
       theme={{
-        useLocalStorage: false,
+        config: {
+          useLocalStorage: false,
+        },
       }}>
       <ColorModeProvider>
         <Button />
@@ -274,7 +280,9 @@ test('initializes mode from prefers-color-scheme media query', () => {
   render(
     <ThemeProvider
       theme={{
-        useColorSchemeMediaQuery: true,
+        config: {
+          useColorSchemeMediaQuery: true,
+        },
       }}>
       <ColorModeProvider>
         <Consumer />
@@ -300,7 +308,9 @@ test('initializes light mode from prefers-color-scheme media query', () => {
   render(
     <ThemeProvider
       theme={{
-        useColorSchemeMediaQuery: true,
+        config: {
+          useColorSchemeMediaQuery: true,
+        },
       }}>
       <ColorModeProvider>
         <Consumer />
@@ -326,7 +336,9 @@ test('does not initialize mode from prefers-color-scheme media query', () => {
   render(
     <ThemeProvider
       theme={{
-        useColorSchemeMediaQuery: true,
+        config: {
+          useColorSchemeMediaQuery: true,
+        },
       }}>
       <ColorModeProvider>
         <Consumer />
@@ -352,7 +364,9 @@ test('does not initialize mode from prefers-color-scheme media query when useCol
   render(
     <ThemeProvider
       theme={{
-        useColorSchemeMediaQuery: false,
+        config: {
+          useColorSchemeMediaQuery: false,
+        },
       }}>
       <ColorModeProvider>
         <Consumer />
@@ -366,7 +380,9 @@ test('ColorModeProvider renders with global colors', () => {
   const root = render(
     <ThemeProvider
       theme={{
-        useCustomProperties: false,
+        config: {
+          useCustomProperties: false,
+        },
         colors: {
           text: 'tomato',
           background: 'black',
@@ -383,7 +399,7 @@ test('ColorModeProvider renders with global colors', () => {
       </ColorModeProvider>
     </ThemeProvider>
   )
-  const style = window.getComputedStyle(root.baseElement)
+  const style = window.getComputedStyle(root.baseElement.parentElement!)
   expect(style.color).toBe('tomato')
   expect(style.backgroundColor).toBe('black')
 })
@@ -414,7 +430,9 @@ test('useThemeUI returns current color mode colors', () => {
     <ThemeProvider
       theme={{
         // minor functional change
-        useCustomProperties: false,
+        config: {
+          useCustomProperties: false,
+        },
         colors: {
           text: 'tomato',
           background: 'black',
@@ -486,7 +504,9 @@ test('warns when initialColorModeName matches a key in theme.colors.modes', () =
   const root = render(
     <ThemeProvider
       theme={{
-        initialColorModeName: 'dark',
+        config: {
+          initialColorModeName: 'dark',
+        },
         colors: {
           text: '#000',
           background: '#fff',
@@ -512,7 +532,9 @@ test('does not warn in production', () => {
   const root = render(
     <ThemeProvider
       theme={{
-        initialColorModeName: 'dark',
+        config: {
+          initialColorModeName: 'dark',
+        },
         colors: {
           text: '#000',
           background: '#fff',
@@ -550,7 +572,9 @@ test('dot notation works with color modes', () => {
   const root = render(
     <ThemeProvider
       theme={{
-        useCustomProperties: false,
+        config: {
+          useCustomProperties: false,
+        },
         colors: {
           header: {
             title: 'blue',
@@ -625,7 +649,9 @@ test('raw color values are passed to theme-ui context when custom properties are
   const root = render(
     <ThemeProvider
       theme={{
-        useColorSchemeMediaQuery: false,
+        config: {
+          useColorSchemeMediaQuery: false,
+        },
         colors: {
           primary: 'tomato',
           modes: {
@@ -747,7 +773,9 @@ test('colorMode accepts function from previous state to new one', () => {
   type MyColorMode = 'serious' | 'cute' | 'hackerman'
 
   const theme: Theme = {
-    initialColorModeName: 'serious',
+    config: {
+      initialColorModeName: 'serious',
+    },
     colors: {
       primary: 'black',
       modes: {