+++ packages/css/src/index.ts
@@ -7,7 +7,6 @@ import {
 } from './types'
 
 export * from './types'
-export { __internalGetUseRootStyles } from './options'
 
 /**
  * Allows for nested scales with shorthand values