+++ packages/theme-provider/src/index.ts
@@ -3,7 +3,6 @@ import {
   useThemeUI,
   ThemeProvider as CoreProvider,
   ThemeProviderProps as CoreThemeProviderProps,
-  __internalGetUseRootStyles,
 } from '@theme-ui/core'
 import { css, Theme } from '@theme-ui/css'
 import { ColorModeProvider } from '@theme-ui/color-modes'
@@ -14,18 +13,20 @@ const RootStyles = () =>
   jsx(Global, {
     styles: (emotionTheme) => {
       const theme = emotionTheme as Theme
-      const use = __internalGetUseRootStyles(theme)
+      const { useRootStyles } = theme.config || theme
 
-      if (use.rootStyles === false || (theme.styles && !theme.styles.root)) {
-        return false
+      if (useRootStyles === false || (theme.styles && !theme.styles.root)) {
+        return null
       }
-      const boxSizing = theme.useBorderBox === false ? undefined : 'border-box'
+
+      const boxSizing =
+        theme.config?.useBorderBox === false ? undefined : 'border-box'
 
       return css({
         '*': {
           boxSizing,
         },
-        [use.scope]: {
+        html: {
           margin: 0,
           variant: 'styles.root',
         },